remove pycliarr, use devopsarr/radarr-py
