+++ pyproject.toml
@@ -5,13 +5,13 @@ description = "continual rename tool for *arr apps"
 authors = [{ name = "Brad Holland", email = "hollanbm@gmail.com" }]
 version = "1"
 dependencies = [
-  "pycliarr>=1.0.26,<2",
   "python-dotenv>=1.0.1,<2",
   "loguru>=0.7.3,<0.8",
   "python-config-parser>=3.1.3,<4",
   "schedule>=1.2.2,<2",
   "sonarr",
   "pydantic>=2,<3",
+  "radarr>=1.2.1",
 ]
 
 [dependency-groups]
@@ -37,9 +37,11 @@ environments = [
 
 [tool.uv.sources]
 sonarr = { git = "https://github.com/devopsarr/sonarr-py.git", tag = "v1.1.1" }
+radarr = { git = "https://github.com/devopsarr/radarr-py.git", tag = "v1.2.1" }
+
 [tool.ty.analysis]
 # These dependencies expose incomplete or overly broad type information.
-replace-imports-with-any = ["pycliarr.**", "schema.**"]
+replace-imports-with-any = ["schema.**"]
 
 [tool.pytest]
 pythonpath = ["./src"]
