+++ src/renamarr/adapter_factory.py
@@ -2,7 +2,7 @@ from enum import StrEnum
 
 from renamarr.protocols import ArrAdapter
 from renamarr.radarr.radarr_adapter import RadarrAdapter
-from renamarr.sonarr_py.sonarr_py_adapter import SonarrPyAdapter
+from renamarr.sonarr.sonarr_adapter import SonarrAdapter
 
 
 class ArrService(StrEnum):
@@ -18,6 +18,6 @@ def create_arr_adapter(service: ArrService, url: str, api_key: str) -> ArrAdapte
         case ArrService.RADARR:
             return RadarrAdapter(url, api_key)
         case ArrService.SONARR:
-            return SonarrPyAdapter(url, api_key)
+            return SonarrAdapter(url, api_key)
         case _:
             raise ValueError(f"Unsupported Arr service: {service!r}")
