+++ tests/test_radarr_adapter_contract.py
@@ -1,9 +1,18 @@
 from unittest.mock import MagicMock, call
 
 import pytest
-from pycliarr.api import RadarrMovieItem
-from pycliarr.api.exceptions import CliArrError
 from pytest_mock import MockerFixture
+from radarr import (
+    CommandResource,
+    MediaManagementConfigResource,
+    MovieEditorResource,
+    MovieResource,
+    RenameMovieResource,
+    RootFolderResource,
+)
+from radarr import CommandResult as RadarrCommandResult
+from radarr import CommandStatus as RadarrCommandStatus
+from radarr.rest import ApiException
 
 from renamarr.exceptions import ArrOperationError
 from renamarr.models.command import CommandStatus
@@ -17,75 +26,99 @@ from renamarr.radarr.radarr_adapter import RadarrAdapter
 
 
 @pytest.fixture
-def radarr_client(mocker: MockerFixture) -> MagicMock:
-    client = mocker.patch("renamarr.radarr.radarr_adapter.RadarrCli").return_value
-    client.api_url_command = "/api/v3/command"
-    return client
+def radarr_apis(mocker: MockerFixture) -> dict[str, MagicMock]:
+    mocker.patch("renamarr.radarr.radarr_adapter.ApiClient")
+    return {
+        name: mocker.patch(f"renamarr.radarr.radarr_adapter.{name}").return_value
+        for name in (
+            "MovieApi",
+            "MediaManagementConfigApi",
+            "CommandApi",
+            "RenameMovieApi",
+            "RootFolderApi",
+            "MovieFolderApi",
+            "MovieEditorApi",
+        )
+    }
 
 
 @pytest.fixture
-def adapter(radarr_client: MagicMock) -> RadarrAdapter:
+def adapter(radarr_apis: dict[str, MagicMock]) -> RadarrAdapter:
     return RadarrAdapter("https://radarr.test", "api-key")
 
 
 def test_maps_movies_to_shared_media_items(
-    adapter: RadarrAdapter, radarr_client: MagicMock
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
 ) -> None:
-    radarr_client.get_movie.return_value = [
-        RadarrMovieItem(id=2, title="Movie B", path="/movies/Movie B"),
-        RadarrMovieItem(id=1, title="Movie A", path="/movies/Movie A"),
+    movie_api = radarr_apis["MovieApi"]
+    movie_api.list_movie.return_value = [
+        MovieResource(id=2, title="Movie B", path="/movies/Movie B"),
+        MovieResource(id=1, title="Movie A", path="/movies/Movie A"),
     ]
 
     assert adapter.list_media_items() == [
         MediaItem(id=2, title="Movie B", path="/movies/Movie B"),
         MediaItem(id=1, title="Movie A", path="/movies/Movie A"),
     ]
-    radarr_client.get_movie.assert_called_once_with()
+    movie_api.list_movie.assert_called_once_with()
 
 
 @pytest.mark.parametrize("enabled", [True, False])
 def test_reads_media_analysis_setting(
-    adapter: RadarrAdapter, radarr_client: MagicMock, enabled: bool
+    adapter: RadarrAdapter,
+    radarr_apis: dict[str, MagicMock],
+    enabled: bool,
 ) -> None:
-    radarr_client.request_get.return_value = {"enableMediaInfo": enabled}
+    media_management_api = radarr_apis["MediaManagementConfigApi"]
+    media_management_api.get_media_management_config.return_value = (
+        MediaManagementConfigResource(enable_media_info=enabled)
+    )
 
     assert adapter.is_media_analysis_enabled() is enabled
-    radarr_client.request_get.assert_called_once_with(
-        path="/api/v3/config/mediamanagement"
-    )
+    media_management_api.get_media_management_config.assert_called_once_with()
 
 
 def test_starts_media_analysis_and_maps_command_status(
-    adapter: RadarrAdapter, radarr_client: MagicMock
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
 ) -> None:
-    radarr_client.request_post.return_value = {"id": 17}
-    radarr_client.get_command.side_effect = [
-        {"status": "started"},
-        {"status": "completed", "result": "failed"},
-        {"status": "completed", "result": "successful"},
+    command_api = radarr_apis["CommandApi"]
+    command_api.create_command.return_value = CommandResource(id=17)
+    command_api.get_command_by_id.side_effect = [
+        CommandResource(status=RadarrCommandStatus.STARTED),
+        CommandResource(
+            status=RadarrCommandStatus.COMPLETED,
+            result=RadarrCommandResult.UNSUCCESSFUL,
+        ),
+        CommandResource(
+            status=RadarrCommandStatus.COMPLETED,
+            result=RadarrCommandResult.SUCCESSFUL,
+        ),
     ]
 
     assert adapter.start_media_analysis() == 17
     assert adapter.get_command_status(17) == CommandStatus(False, False)
     assert adapter.get_command_status(17) == CommandStatus(True, False)
     assert adapter.get_command_status(17) == CommandStatus(True, True)
-    radarr_client.request_post.assert_called_once_with(
-        "/api/v3/command",
-        json_data={"name": "RescanMovie", "priority": "high"},
-    )
-    radarr_client.get_command.assert_has_calls(
-        [call(cid=17), call(cid=17), call(cid=17)]
-    )
+    command = command_api.create_command.call_args.args[0]
+    assert command.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "name": "RescanMovie",
+        "priority": "high",
+    }
+    command_api.get_command_by_id.assert_has_calls([call(17), call(17), call(17)])
 
 
 def test_maps_and_combines_file_rename_previews(
-    adapter: RadarrAdapter, radarr_client: MagicMock
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
 ) -> None:
     movie_a = MediaItem(1, "Movie A", "/movies/Movie A")
     movie_b = MediaItem(2, "Movie B", "/movies/Movie B")
-    radarr_client.request_get.side_effect = [
+    rename_api = radarr_apis["RenameMovieApi"]
+    rename_api.list_rename.side_effect = [
         [],
-        [{"movieFileId": 10}, {"movieFileId": 11}],
+        [
+            RenameMovieResource(movie_file_id=10),
+            RenameMovieResource(movie_file_id=11),
+        ],
     ]
 
     assert adapter.get_file_rename_candidate(movie_a) is None
@@ -97,84 +130,94 @@ def test_maps_and_combines_file_rename_previews(
     assert adapter.build_file_rename_batches((candidate_a, candidate_b)) == [
         FileRenameBatch((1, 2), (10, 11, 20), "Movie A, Movie B")
     ]
-    radarr_client.request_get.assert_has_calls(
-        [
-            call(path="/api/v3/rename", url_params={"movieId": 1}),
-            call(path="/api/v3/rename", url_params={"movieId": 1}),
-        ]
-    )
+    rename_api.list_rename.assert_has_calls([call(movie_id=[1]), call(movie_id=[1])])
 
 
 def test_starts_aggregate_movie_rename(
-    adapter: RadarrAdapter, radarr_client: MagicMock
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
 ) -> None:
     batch = FileRenameBatch((1, 2), (10, 20), "Movie A, Movie B")
-    radarr_client.request_post.return_value = {"id": 23}
+    command_api = radarr_apis["CommandApi"]
+    command_api.create_command.return_value = CommandResource(id=23)
 
     assert adapter.start_file_rename(batch) == 23
-    radarr_client.request_post.assert_called_once_with(
-        "/api/v3/command",
-        json_data={"name": "RenameMovie", "movieIds": [1, 2]},
-    )
+    command = command_api.create_command.call_args.args[0]
+    assert command.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "name": "RenameMovie",
+        "movieIds": [1, 2],
+    }
 
 
 def test_uses_radarr_folder_endpoints_and_payloads(
-    adapter: RadarrAdapter, radarr_client: MagicMock
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
 ) -> None:
     movie_a = MediaItem(1, "Movie A", "/movies/Old A")
     movie_b = MediaItem(2, "Movie B", "/movies/Old B")
     batch = FolderRenameBatch("/movies", (movie_a, movie_b), False)
-    radarr_client.get_root_folder.return_value = [
-        {"path": "/movies"},
-        {"path": "/movies-4k"},
+    root_folder_api = radarr_apis["RootFolderApi"]
+    root_folder_api.list_root_folder.return_value = [
+        RootFolderResource(path="/movies"),
+        RootFolderResource(path="/movies-4k"),
     ]
-    radarr_client.request_get.return_value = {"folder": "Movie A (2026)"}
-    radarr_client.request_put.return_value = {}
-    radarr_client.request_post.return_value = {"id": 29}
+    movie_folder_api = radarr_apis["MovieFolderApi"]
+    folder_response = MagicMock(status=200)
+    folder_response.read.return_value = b'{"folder": "Movie A (2026)"}'
+    movie_folder_api.get_movie_folder_without_preload_content.return_value = (
+        folder_response
+    )
+    command_api = radarr_apis["CommandApi"]
+    command_api.create_command.return_value = CommandResource(id=29)
 
     assert adapter.list_root_folders() == ["/movies", "/movies-4k"]
     assert adapter.get_expected_folder_name(movie_a) == "Movie A (2026)"
     assert adapter.move_folder(batch) is None
     assert adapter.start_folder_rescan(batch) == 29
-    radarr_client.request_get.assert_called_once_with(path="/api/v3/movie/1/folder")
-    radarr_client.request_put.assert_called_once_with(
-        path="/api/v3/movie/editor",
-        json_data={
-            "rootFolderPath": "/movies",
-            "movieIds": [1, 2],
-            "moveFiles": False,
-        },
+    movie_folder_api.get_movie_folder_without_preload_content.assert_called_once_with(
+        id=1
     )
-    radarr_client.request_post.assert_called_once_with(
-        "/api/v3/command",
-        json_data={
-            "priority": "high",
-            "name": "RefreshMovie",
-            "movieIds": [1, 2],
-        },
-    )
+    folder_response.read.assert_called_once_with()
+    folder_response.release_conn.assert_called_once_with()
+    movie_editor_api = radarr_apis["MovieEditorApi"]
+    editor = movie_editor_api.put_movie_editor.call_args.args[0]
+    assert isinstance(editor, MovieEditorResource)
+    assert editor.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "movieIds": [1, 2],
+        "rootFolderPath": "/movies",
+        "moveFiles": False,
+    }
+    command = command_api.create_command.call_args.args[0]
+    assert command.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "name": "RefreshMovie",
+        "priority": "high",
+        "movieIds": [1, 2],
+    }
 
 
 @pytest.mark.parametrize(
-    ("boundary", "client_method"),
+    ("boundary", "api_name", "method_name"),
     [
-        ("list", "get_movie"),
-        ("setting", "request_get"),
-        ("analysis", "request_post"),
-        ("status", "get_command"),
-        ("preview", "request_get"),
-        ("rename", "request_post"),
-        ("roots", "get_root_folder"),
-        ("folder", "request_get"),
-        ("move", "request_put"),
-        ("rescan", "request_post"),
+        ("list", "MovieApi", "list_movie"),
+        ("setting", "MediaManagementConfigApi", "get_media_management_config"),
+        ("analysis", "CommandApi", "create_command"),
+        ("status", "CommandApi", "get_command_by_id"),
+        ("preview", "RenameMovieApi", "list_rename"),
+        ("rename", "CommandApi", "create_command"),
+        ("roots", "RootFolderApi", "list_root_folder"),
+        (
+            "folder",
+            "MovieFolderApi",
+            "get_movie_folder_without_preload_content",
+        ),
+        ("move", "MovieEditorApi", "put_movie_editor"),
+        ("rescan", "CommandApi", "create_command"),
     ],
 )
-def test_translates_cliarr_errors_at_every_api_boundary(
+def test_translates_api_errors_at_every_boundary(
     adapter: RadarrAdapter,
-    radarr_client: MagicMock,
+    radarr_apis: dict[str, MagicMock],
     boundary: str,
-    client_method: str,
+    api_name: str,
+    method_name: str,
 ) -> None:
     item = MediaItem(1, "Movie", "/movies/Movie")
     file_batch = FileRenameBatch((1,), (10,), "Movie")
@@ -191,18 +234,21 @@ def test_translates_cliarr_errors_at_every_api_boundary(
         "move": lambda: adapter.move_folder(folder_batch),
         "rescan": lambda: adapter.start_folder_rescan(folder_batch),
     }
-    getattr(radarr_client, client_method).side_effect = CliArrError("broken")
+    getattr(radarr_apis[api_name], method_name).side_effect = ApiException(
+        reason="broken"
+    )
 
-    with pytest.raises(ArrOperationError, match="failed: broken") as error:
+    with pytest.raises(ArrOperationError, match="failed:") as error:
         operations[boundary]()
 
-    assert isinstance(error.value.__cause__, CliArrError)
+    assert isinstance(error.value.__cause__, ApiException)
+    assert "broken" in str(error.value)
 
 
 def test_does_not_translate_unexpected_errors(
-    adapter: RadarrAdapter, radarr_client: MagicMock
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
 ) -> None:
-    radarr_client.get_movie.side_effect = RuntimeError("programming error")
+    radarr_apis["MovieApi"].list_movie.side_effect = RuntimeError("programming error")
 
     with pytest.raises(RuntimeError, match="programming error"):
         adapter.list_media_items()
@@ -211,49 +257,42 @@ def test_does_not_translate_unexpected_errors(
 @pytest.mark.parametrize(
     ("operation", "response", "message"),
     [
-        ("list", RadarrMovieItem(id=1), "Expected a list of movies"),
-        ("setting", [], "Expected an object response"),
-        ("setting", {}, "Expected enableMediaInfo"),
-        ("setting", {"enableMediaInfo": 1}, "Expected enableMediaInfo"),
-        ("analysis", {}, "Expected a numeric command ID"),
-        ("analysis", {"id": "1"}, "Expected a numeric command ID"),
-        ("preview", {}, "Expected a list of rename previews"),
-        ("preview", [[]], "Expected an object response"),
+        ("list", MovieResource(id=1), "Expected a list of movies"),
         (
-            "preview",
-            [{"movieFileId": "10"}],
-            "Expected a movie file ID",
+            "setting",
+            MediaManagementConfigResource(),
+            "Expected enableMediaInfo",
         ),
+        ("analysis", CommandResource(), "Expected a numeric command ID"),
+        ("preview", {}, "Expected a list of rename previews"),
         ("roots", {}, "Expected a list of root folders"),
-        ("roots", [[]], "Expected an object response"),
-        ("roots", [{"path": 1}], "Expected a root-folder path"),
-        ("folder", {}, "Expected a folder name"),
-        ("folder", {"folder": 1}, "Expected a folder name"),
     ],
 )
-def test_rejects_malformed_radarr_responses(
+def test_rejects_malformed_top_level_responses(
     adapter: RadarrAdapter,
-    radarr_client: MagicMock,
+    radarr_apis: dict[str, MagicMock],
     operation: str,
     response: object,
     message: str,
 ) -> None:
     item = MediaItem(1, "Movie", "/movies/Movie")
     operations = {
-        "list": (radarr_client.get_movie, adapter.list_media_items),
+        "list": (radarr_apis["MovieApi"].list_movie, adapter.list_media_items),
         "setting": (
-            radarr_client.request_get,
+            radarr_apis["MediaManagementConfigApi"].get_media_management_config,
             adapter.is_media_analysis_enabled,
         ),
-        "analysis": (radarr_client.request_post, adapter.start_media_analysis),
+        "analysis": (
+            radarr_apis["CommandApi"].create_command,
+            adapter.start_media_analysis,
+        ),
         "preview": (
-            radarr_client.request_get,
+            radarr_apis["RenameMovieApi"].list_rename,
             lambda: adapter.get_file_rename_candidate(item),
         ),
-        "roots": (radarr_client.get_root_folder, adapter.list_root_folders),
-        "folder": (
-            radarr_client.request_get,
-            lambda: adapter.get_expected_folder_name(item),
+        "roots": (
+            radarr_apis["RootFolderApi"].list_root_folder,
+            adapter.list_root_folders,
         ),
     }
     client_call, adapter_call = operations[operation]
@@ -261,3 +300,89 @@ def test_rejects_malformed_radarr_responses(
 
     with pytest.raises(TypeError, match=message):
         adapter_call()
+
+
+@pytest.mark.parametrize(
+    ("movie", "message"),
+    [
+        (MovieResource(title="Movie", path="/movies/Movie"), "Expected a movie ID"),
+        (MovieResource(id=1, path="/movies/Movie"), "Expected a movie title"),
+        (MovieResource(id=1, title="Movie"), "Expected a movie path"),
+    ],
+)
+def test_rejects_movies_missing_required_fields(
+    adapter: RadarrAdapter,
+    radarr_apis: dict[str, MagicMock],
+    movie: MovieResource,
+    message: str,
+) -> None:
+    radarr_apis["MovieApi"].list_movie.return_value = [movie]
+
+    with pytest.raises(TypeError, match=message):
+        adapter.list_media_items()
+
+
+def test_rejects_rename_preview_missing_movie_file_id(
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
+) -> None:
+    radarr_apis["RenameMovieApi"].list_rename.return_value = [RenameMovieResource()]
+
+    with pytest.raises(TypeError, match="Expected a movie file ID"):
+        adapter.get_file_rename_candidate(MediaItem(1, "Movie", "/movies/Movie"))
+
+
+def test_rejects_root_folder_missing_path(
+    adapter: RadarrAdapter, radarr_apis: dict[str, MagicMock]
+) -> None:
+    radarr_apis["RootFolderApi"].list_root_folder.return_value = [RootFolderResource()]
+
+    with pytest.raises(TypeError, match="Expected a root-folder path"):
+        adapter.list_root_folders()
+
+
+@pytest.mark.parametrize("status", [199, 300])
+def test_rejects_non_successful_folder_response_before_parsing(
+    adapter: RadarrAdapter,
+    radarr_apis: dict[str, MagicMock],
+    status: int,
+) -> None:
+    folder_response = MagicMock(status=status, reason="Unexpected status")
+    folder_response.read.return_value = b"not json"
+    radarr_apis[
+        "MovieFolderApi"
+    ].get_movie_folder_without_preload_content.return_value = folder_response
+
+    with pytest.raises(
+        ArrOperationError, match="Resolve Radarr folder.*failed"
+    ) as error:
+        adapter.get_expected_folder_name(MediaItem(1, "Movie", "/movies/Movie"))
+
+    assert isinstance(error.value.__cause__, ApiException)
+    assert error.value.__cause__.status == status
+    assert error.value.__cause__.body == "not json"
+    folder_response.release_conn.assert_called_once_with()
+
+
+@pytest.mark.parametrize(
+    ("payload", "message"),
+    [
+        (b"[]", "Expected an object response"),
+        (b'{"folder": 1}', "Expected a folder name"),
+    ],
+)
+def test_rejects_malformed_folder_response(
+    adapter: RadarrAdapter,
+    radarr_apis: dict[str, MagicMock],
+    payload: bytes,
+    message: str,
+) -> None:
+    folder_response = MagicMock(status=200)
+    folder_response.read.return_value = payload
+    radarr_apis[
+        "MovieFolderApi"
+    ].get_movie_folder_without_preload_content.return_value = folder_response
+
+    with pytest.raises(TypeError, match=message):
+        adapter.get_expected_folder_name(MediaItem(1, "Movie", "/movies/Movie"))
+
+    folder_response.release_conn.assert_called_once_with()
