+++ tests/test_sonarr_adapter_contract.py
@@ -1,9 +1,18 @@
 from unittest.mock import MagicMock, call
 
 import pytest
-from pycliarr.api import SonarrSerieItem
-from pycliarr.api.exceptions import CliArrError
 from pytest_mock import MockerFixture
+from sonarr import (
+    CommandResource,
+    MediaManagementConfigResource,
+    RenameEpisodeResource,
+    RootFolderResource,
+    SeriesEditorResource,
+    SeriesResource,
+)
+from sonarr import CommandResult as SonarrCommandResult
+from sonarr import CommandStatus as SonarrCommandStatus
+from sonarr.rest import ApiException
 
 from renamarr.exceptions import ArrOperationError
 from renamarr.models.command import CommandStatus
@@ -17,83 +26,106 @@ from renamarr.sonarr.sonarr_adapter import SonarrAdapter
 
 
 @pytest.fixture
-def sonarr_client(mocker: MockerFixture) -> MagicMock:
-    client = mocker.patch(
-        "renamarr.sonarr.sonarr_adapter.SonarrCli", autospec=True
-    ).return_value
-    client.api_url_command = "/api/v3/command"
-    return client
+def sonarr_apis(mocker: MockerFixture) -> dict[str, MagicMock]:
+    mocker.patch("renamarr.sonarr.sonarr_adapter.ApiClient")
+    return {
+        name: mocker.patch(f"renamarr.sonarr.sonarr_adapter.{name}").return_value
+        for name in (
+            "SeriesApi",
+            "MediaManagementConfigApi",
+            "CommandApi",
+            "RenameEpisodeApi",
+            "RootFolderApi",
+            "SeriesFolderApi",
+            "SeriesEditorApi",
+        )
+    }
 
 
 @pytest.fixture
-def adapter(sonarr_client: MagicMock) -> SonarrAdapter:
+def adapter(sonarr_apis: dict[str, MagicMock]) -> SonarrAdapter:
     return SonarrAdapter("https://sonarr.test", "api-key")
 
 
 def test_maps_series_to_shared_media_items(
-    adapter: SonarrAdapter, sonarr_client: MagicMock
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
 ) -> None:
-    sonarr_client.get_serie.return_value = [
-        SonarrSerieItem(id=2, title="Show B", path="/tv/Show B"),
-        SonarrSerieItem(id=1, title="Show A", path="/tv/Show A"),
+    series_api = sonarr_apis["SeriesApi"]
+    series_api.list_series.return_value = [
+        SeriesResource(id=2, title="Show B", path="/tv/Show B"),
+        SeriesResource(id=1, title="Show A", path="/tv/Show A"),
     ]
 
     assert adapter.list_media_items() == [
         MediaItem(id=2, title="Show B", path="/tv/Show B"),
         MediaItem(id=1, title="Show A", path="/tv/Show A"),
     ]
-    sonarr_client.get_serie.assert_called_once_with()
+    series_api.list_series.assert_called_once_with()
 
 
 @pytest.mark.parametrize("enabled", [True, False])
 def test_reads_media_analysis_setting(
-    adapter: SonarrAdapter, sonarr_client: MagicMock, enabled: bool
+    adapter: SonarrAdapter,
+    sonarr_apis: dict[str, MagicMock],
+    enabled: bool,
 ) -> None:
-    sonarr_client.request_get.return_value = {"enableMediaInfo": enabled}
+    media_management_api = sonarr_apis["MediaManagementConfigApi"]
+    media_management_api.get_media_management_config.return_value = (
+        MediaManagementConfigResource(enable_media_info=enabled)
+    )
 
     assert adapter.is_media_analysis_enabled() is enabled
-    sonarr_client.request_get.assert_called_once_with(
-        path="/api/v3/config/mediamanagement"
-    )
+    media_management_api.get_media_management_config.assert_called_once_with()
 
 
 def test_starts_media_analysis_and_maps_command_status(
-    adapter: SonarrAdapter, sonarr_client: MagicMock
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
 ) -> None:
-    sonarr_client.request_post.return_value = {"id": 17}
-    sonarr_client.get_command.side_effect = [
-        {"status": "started"},
-        {"status": "completed", "result": "failed"},
-        {"status": "completed", "result": "successful"},
+    command_api = sonarr_apis["CommandApi"]
+    command_api.create_command.return_value = CommandResource(id=17)
+    command_api.get_command_by_id.side_effect = [
+        CommandResource(status=SonarrCommandStatus.STARTED),
+        CommandResource(
+            status=SonarrCommandStatus.COMPLETED,
+            result=SonarrCommandResult.UNSUCCESSFUL,
+        ),
+        CommandResource(
+            status=SonarrCommandStatus.COMPLETED,
+            result=SonarrCommandResult.SUCCESSFUL,
+        ),
     ]
 
     assert adapter.start_media_analysis() == 17
     assert adapter.get_command_status(17) == CommandStatus(False, False)
     assert adapter.get_command_status(17) == CommandStatus(True, False)
     assert adapter.get_command_status(17) == CommandStatus(True, True)
-    sonarr_client.request_post.assert_called_once_with(
-        "/api/v3/command",
-        json_data={"name": "RescanSeries", "priority": "high"},
-    )
-    sonarr_client.get_command.assert_has_calls(
-        [call(cid=17), call(cid=17), call(cid=17)]
-    )
+    command = command_api.create_command.call_args.args[0]
+    assert command.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "name": "RescanSeries",
+        "priority": "high",
+    }
+    command_api.get_command_by_id.assert_has_calls([call(17), call(17), call(17)])
 
 
 def test_maps_episode_previews_and_builds_one_batch_per_series(
-    adapter: SonarrAdapter, sonarr_client: MagicMock
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
 ) -> None:
     show_a = MediaItem(1, "Show A", "/tv/Show A")
     show_b = MediaItem(2, "Show B", "/tv/Show B")
-    sonarr_client.request_get.side_effect = [
+    rename_api = sonarr_apis["RenameEpisodeApi"]
+    rename_api.list_rename.side_effect = [
         [],
         [
-            {"episodeFileId": 10, "seasonNumber": 1, "episodeNumbers": [1]},
-            {
-                "episodeFileId": 20,
-                "seasonNumber": 2,
-                "episodeNumbers": [3, 4],
-            },
+            RenameEpisodeResource(
+                episode_file_id=10,
+                season_number=1,
+                episode_numbers=[1],
+            ),
+            RenameEpisodeResource(
+                episode_file_id=20,
+                season_number=2,
+                episode_numbers=[3, 4],
+            ),
         ],
     ]
 
@@ -108,81 +140,95 @@ def test_maps_episode_previews_and_builds_one_batch_per_series(
         FileRenameBatch((2,), (30,), "Show B: S03E05"),
     ]
     assert adapter.build_file_rename_batches(()) == []
-    sonarr_client.request_get.assert_has_calls(
-        [
-            call(path="/api/v3/rename", url_params={"seriesId": 1}),
-            call(path="/api/v3/rename", url_params={"seriesId": 1}),
-        ]
-    )
+    rename_api.list_rename.assert_has_calls([call(series_id=1), call(series_id=1)])
 
 
 def test_starts_series_file_rename(
-    adapter: SonarrAdapter, sonarr_client: MagicMock
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
 ) -> None:
     batch = FileRenameBatch((7,), (10, 20), "S01E01, S01E02")
-    sonarr_client.rename_files.return_value = {"id": 23}
+    command_api = sonarr_apis["CommandApi"]
+    command_api.create_command.return_value = CommandResource(id=23)
 
     assert adapter.start_file_rename(batch) == 23
-    sonarr_client.rename_files.assert_called_once_with([10, 20], 7)
+    command = command_api.create_command.call_args.args[0]
+    assert command.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "name": "RenameFiles",
+        "files": [10, 20],
+        "seriesId": 7,
+    }
 
 
 def test_uses_sonarr_folder_endpoints_and_payloads(
-    adapter: SonarrAdapter, sonarr_client: MagicMock
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
 ) -> None:
     show_a = MediaItem(1, "Show A", "/tv/Old A")
     show_b = MediaItem(2, "Show B", "/tv/Old B")
     batch = FolderRenameBatch("/tv", (show_a, show_b), False)
-    sonarr_client.get_root_folder.return_value = [
-        {"path": "/tv"},
-        {"path": "/tv-anime"},
+    root_folder_api = sonarr_apis["RootFolderApi"]
+    root_folder_api.list_root_folder.return_value = [
+        RootFolderResource(path="/tv"),
+        RootFolderResource(path="/tv-anime"),
     ]
-    sonarr_client.request_get.return_value = {"folder": "Show A (2026)"}
-    sonarr_client.request_put.return_value = {}
-    sonarr_client.request_post.return_value = {"id": 29}
+    series_folder_api = sonarr_apis["SeriesFolderApi"]
+    folder_response = MagicMock(status=200)
+    folder_response.read.return_value = b'{"folder": "Show A (2026)"}'
+    series_folder_api.get_series_folder_without_preload_content.return_value = (
+        folder_response
+    )
+    command_api = sonarr_apis["CommandApi"]
+    command_api.create_command.return_value = CommandResource(id=29)
 
     assert adapter.list_root_folders() == ["/tv", "/tv-anime"]
     assert adapter.get_expected_folder_name(show_a) == "Show A (2026)"
     assert adapter.move_folder(batch) is None
     assert adapter.start_folder_rescan(batch) == 29
-    sonarr_client.request_get.assert_called_once_with(path="/api/v3/series/1/folder")
-    sonarr_client.request_put.assert_called_once_with(
-        path="/api/v3/series/editor",
-        json_data={
-            "rootFolderPath": "/tv",
-            "seriesIds": [1, 2],
-            "moveFiles": False,
-        },
+    series_folder_api.get_series_folder_without_preload_content.assert_called_once_with(
+        id=1
     )
-    sonarr_client.request_post.assert_called_once_with(
-        "/api/v3/command",
-        json_data={
-            "name": "RescanSeries",
-            "priority": "high",
-            "seriesIds": [1, 2],
-        },
-    )
+    folder_response.read.assert_called_once_with()
+    folder_response.release_conn.assert_called_once_with()
+    series_editor_api = sonarr_apis["SeriesEditorApi"]
+    editor = series_editor_api.put_series_editor.call_args.args[0]
+    assert isinstance(editor, SeriesEditorResource)
+    assert editor.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "seriesIds": [1, 2],
+        "rootFolderPath": "/tv",
+        "moveFiles": False,
+    }
+    command = command_api.create_command.call_args.args[0]
+    assert command.model_dump(mode="json", by_alias=True, exclude_none=True) == {
+        "name": "RescanSeries",
+        "priority": "high",
+        "seriesIds": [1, 2],
+    }
 
 
 @pytest.mark.parametrize(
-    ("boundary", "client_method"),
+    ("boundary", "api_name", "method_name"),
     [
-        ("list", "get_serie"),
-        ("setting", "request_get"),
-        ("analysis", "request_post"),
-        ("status", "get_command"),
-        ("preview", "request_get"),
-        ("rename", "rename_files"),
-        ("roots", "get_root_folder"),
-        ("folder", "request_get"),
-        ("move", "request_put"),
-        ("rescan", "request_post"),
+        ("list", "SeriesApi", "list_series"),
+        ("setting", "MediaManagementConfigApi", "get_media_management_config"),
+        ("analysis", "CommandApi", "create_command"),
+        ("status", "CommandApi", "get_command_by_id"),
+        ("preview", "RenameEpisodeApi", "list_rename"),
+        ("rename", "CommandApi", "create_command"),
+        ("roots", "RootFolderApi", "list_root_folder"),
+        (
+            "folder",
+            "SeriesFolderApi",
+            "get_series_folder_without_preload_content",
+        ),
+        ("move", "SeriesEditorApi", "put_series_editor"),
+        ("rescan", "CommandApi", "create_command"),
     ],
 )
-def test_translates_cliarr_errors_at_every_api_boundary(
+def test_translates_api_errors_at_every_boundary(
     adapter: SonarrAdapter,
-    sonarr_client: MagicMock,
+    sonarr_apis: dict[str, MagicMock],
     boundary: str,
-    client_method: str,
+    api_name: str,
+    method_name: str,
 ) -> None:
     item = MediaItem(1, "Show", "/tv/Show")
     file_batch = FileRenameBatch((1,), (10,), "S01E01")
@@ -199,18 +245,21 @@ def test_translates_cliarr_errors_at_every_api_boundary(
         "move": lambda: adapter.move_folder(folder_batch),
         "rescan": lambda: adapter.start_folder_rescan(folder_batch),
     }
-    getattr(sonarr_client, client_method).side_effect = CliArrError("broken")
+    getattr(sonarr_apis[api_name], method_name).side_effect = ApiException(
+        reason="broken"
+    )
 
-    with pytest.raises(ArrOperationError, match="failed: broken") as error:
+    with pytest.raises(ArrOperationError, match="failed:") as error:
         operations[boundary]()
 
-    assert isinstance(error.value.__cause__, CliArrError)
+    assert isinstance(error.value.__cause__, ApiException)
+    assert "broken" in str(error.value)
 
 
 def test_does_not_translate_unexpected_errors(
-    adapter: SonarrAdapter, sonarr_client: MagicMock
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
 ) -> None:
-    sonarr_client.get_serie.side_effect = RuntimeError("programming error")
+    sonarr_apis["SeriesApi"].list_series.side_effect = RuntimeError("programming error")
 
     with pytest.raises(RuntimeError, match="programming error"):
         adapter.list_media_items()
@@ -219,88 +268,42 @@ def test_does_not_translate_unexpected_errors(
 @pytest.mark.parametrize(
     ("operation", "response", "message"),
     [
-        ("list", SonarrSerieItem(id=1), "Expected a list of series"),
-        ("setting", [], "Expected an object response"),
-        ("setting", {}, "Expected enableMediaInfo"),
-        ("setting", {"enableMediaInfo": 1}, "Expected enableMediaInfo"),
-        ("analysis", {}, "Expected a numeric command ID"),
-        ("analysis", {"id": "1"}, "Expected a numeric command ID"),
+        ("list", SeriesResource(id=1), "Expected a list of series"),
+        (
+            "setting",
+            MediaManagementConfigResource(),
+            "Expected enableMediaInfo",
+        ),
+        ("analysis", CommandResource(), "Expected a numeric command ID"),
         ("preview", {}, "Expected a list of rename previews"),
-        ("preview", [[]], "Expected an object response"),
-        (
-            "preview",
-            [
-                {
-                    "episodeFileId": "10",
-                    "seasonNumber": 1,
-                    "episodeNumbers": [1],
-                }
-            ],
-            "Expected an episode file ID",
-        ),
-        (
-            "preview",
-            [
-                {
-                    "episodeFileId": 10,
-                    "seasonNumber": "1",
-                    "episodeNumbers": [1],
-                }
-            ],
-            "Expected a season number",
-        ),
-        (
-            "preview",
-            [
-                {
-                    "episodeFileId": 10,
-                    "seasonNumber": 1,
-                    "episodeNumbers": 1,
-                }
-            ],
-            "Expected episode numbers",
-        ),
-        (
-            "preview",
-            [
-                {
-                    "episodeFileId": 10,
-                    "seasonNumber": 1,
-                    "episodeNumbers": ["1"],
-                }
-            ],
-            "Expected integer episode numbers",
-        ),
         ("roots", {}, "Expected a list of root folders"),
-        ("roots", [[]], "Expected an object response"),
-        ("roots", [{"path": 1}], "Expected a root-folder path"),
-        ("folder", {}, "Expected a folder name"),
-        ("folder", {"folder": 1}, "Expected a folder name"),
     ],
 )
-def test_rejects_malformed_sonarr_responses(
+def test_rejects_malformed_top_level_responses(
     adapter: SonarrAdapter,
-    sonarr_client: MagicMock,
+    sonarr_apis: dict[str, MagicMock],
     operation: str,
     response: object,
     message: str,
 ) -> None:
     item = MediaItem(1, "Show", "/tv/Show")
     operations = {
-        "list": (sonarr_client.get_serie, adapter.list_media_items),
+        "list": (sonarr_apis["SeriesApi"].list_series, adapter.list_media_items),
         "setting": (
-            sonarr_client.request_get,
+            sonarr_apis["MediaManagementConfigApi"].get_media_management_config,
             adapter.is_media_analysis_enabled,
         ),
-        "analysis": (sonarr_client.request_post, adapter.start_media_analysis),
+        "analysis": (
+            sonarr_apis["CommandApi"].create_command,
+            adapter.start_media_analysis,
+        ),
         "preview": (
-            sonarr_client.request_get,
+            sonarr_apis["RenameEpisodeApi"].list_rename,
             lambda: adapter.get_file_rename_candidate(item),
         ),
-        "roots": (sonarr_client.get_root_folder, adapter.list_root_folders),
-        "folder": (
-            sonarr_client.request_get,
-            lambda: adapter.get_expected_folder_name(item),
+        "roots": (
+            sonarr_apis["RootFolderApi"].list_root_folder,
+            adapter.list_root_folders,
         ),
     }
     client_call, adapter_call = operations[operation]
@@ -308,3 +311,109 @@ def test_rejects_malformed_sonarr_responses(
 
     with pytest.raises(TypeError, match=message):
         adapter_call()
+
+
+@pytest.mark.parametrize(
+    ("series", "message"),
+    [
+        (SeriesResource(title="Show", path="/tv/Show"), "Expected a series ID"),
+        (SeriesResource(id=1, path="/tv/Show"), "Expected a series title"),
+        (SeriesResource(id=1, title="Show"), "Expected a series path"),
+    ],
+)
+def test_rejects_series_missing_required_fields(
+    adapter: SonarrAdapter,
+    sonarr_apis: dict[str, MagicMock],
+    series: SeriesResource,
+    message: str,
+) -> None:
+    sonarr_apis["SeriesApi"].list_series.return_value = [series]
+
+    with pytest.raises(TypeError, match=message):
+        adapter.list_media_items()
+
+
+@pytest.mark.parametrize(
+    ("preview", "message"),
+    [
+        (
+            RenameEpisodeResource(season_number=1, episode_numbers=[1]),
+            "Expected an episode file ID",
+        ),
+        (
+            RenameEpisodeResource(episode_file_id=10, episode_numbers=[1]),
+            "Expected a season number",
+        ),
+        (
+            RenameEpisodeResource(episode_file_id=10, season_number=1),
+            "Expected episode numbers",
+        ),
+    ],
+)
+def test_rejects_rename_preview_missing_required_fields(
+    adapter: SonarrAdapter,
+    sonarr_apis: dict[str, MagicMock],
+    preview: RenameEpisodeResource,
+    message: str,
+) -> None:
+    sonarr_apis["RenameEpisodeApi"].list_rename.return_value = [preview]
+
+    with pytest.raises(TypeError, match=message):
+        adapter.get_file_rename_candidate(MediaItem(1, "Show", "/tv/Show"))
+
+
+def test_rejects_root_folder_missing_path(
+    adapter: SonarrAdapter, sonarr_apis: dict[str, MagicMock]
+) -> None:
+    sonarr_apis["RootFolderApi"].list_root_folder.return_value = [RootFolderResource()]
+
+    with pytest.raises(TypeError, match="Expected a root-folder path"):
+        adapter.list_root_folders()
+
+
+@pytest.mark.parametrize("status", [199, 300])
+def test_rejects_non_successful_folder_response_before_parsing(
+    adapter: SonarrAdapter,
+    sonarr_apis: dict[str, MagicMock],
+    status: int,
+) -> None:
+    folder_response = MagicMock(status=status, reason="Unexpected status")
+    folder_response.read.return_value = b"not json"
+    sonarr_apis[
+        "SeriesFolderApi"
+    ].get_series_folder_without_preload_content.return_value = folder_response
+
+    with pytest.raises(
+        ArrOperationError, match="Resolve Sonarr folder.*failed"
+    ) as error:
+        adapter.get_expected_folder_name(MediaItem(1, "Show", "/tv/Show"))
+
+    assert isinstance(error.value.__cause__, ApiException)
+    assert error.value.__cause__.status == status
+    assert error.value.__cause__.body == "not json"
+    folder_response.release_conn.assert_called_once_with()
+
+
+@pytest.mark.parametrize(
+    ("payload", "message"),
+    [
+        (b"[]", "Expected an object response"),
+        (b'{"folder": 1}', "Expected a folder name"),
+    ],
+)
+def test_rejects_malformed_folder_response(
+    adapter: SonarrAdapter,
+    sonarr_apis: dict[str, MagicMock],
+    payload: bytes,
+    message: str,
+) -> None:
+    folder_response = MagicMock(status=200)
+    folder_response.read.return_value = payload
+    sonarr_apis[
+        "SeriesFolderApi"
+    ].get_series_folder_without_preload_content.return_value = folder_response
+
+    with pytest.raises(TypeError, match=message):
+        adapter.get_expected_folder_name(MediaItem(1, "Show", "/tv/Show"))
+
+    folder_response.release_conn.assert_called_once_with()
